+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Server hardware
 - Server hardware type
 - Server profile
+- Server profile template
 - Uplink set
 
 # v2.3.0 (2018-06-26)